requirements.txt
llama-index==0.10.0
llama-index-embeddings-openai==0.1.7
llama-index-llms-openai==0.1.15
openai==1.12.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tiktoken>=0.4.0
nltk>=3.8
transformers>=4.30.0
torch>=1.13.0
sentence-transformers>=2.2.0

# YAML 配置支持
PyYAML>=6.0

# 向量数据库支持
chromadb>=0.4.15
llama-index-vector-stores-chroma>=0.1.4

# PDF 处理相关包
llama-index-readers-file==0.1.4
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.5
pypdf>=3.0.0
unstructured>=0.10.0
pytesseract>=0.3.10
pillow>=9.0.0 